TODO, :Ls, :Lsnew
lots added to the TODO list

:Ls and :Lsnew were implemented. To accomodate this, lsbuffer#new() and
lsbuffer#open() had to be changed.

Before, the function declaration was lsbuffer#new(sp, ...) where a:sp was 'e',
'v', or '' for 'enew', 'vnew', or 'new' (resp.). Now, it is
lsbuffer#new(split=1, mods='', cwd='') where a:split, a:mods, and a:cwd are
boolean indicated whether a split should occur, the string of <mods> passed (can
be set manually as a string by passing 'vertical aboveleft', for instance), and
the current working directory. Any can be omitted, and they will default to
v:true, '', and getcwd().

Before, the function declaraction for the other function was lsbuffer#open(sp)
similar to lsbuffer#new(..). Now, it is lsbuffer#open(split=1, mods='') where
a:split is a boolean indicating if it should split and mods is a string of
space-separated mods.

There seems to be an issue when deleting multipls files. I'd also like to have
one confirm for groups of files and one double confirm for non-empty
directories.

Also, change someo 1s and 0s to v:true and v:false.