Improves cursor restore; much better plugin file
Now when a directory is changed, the name of the file under the cursor
is stored in `b:lslinenrs` (name change pending) instead of the line
number. This is because if the listing changes while in another
directory, then the line number will end up being wrong.

Also, better information is stored. So, when you navigate backwards to
the parent directory (with h) for the first time, it uses the name of
the previous directory you were in to automatically jump to the right
row. This help navigation a lot. If you call `lsbuffer#newcwd('../..')`, I
think this will break, though. **TODO**

Filtering now happens on buffer- and global-levels using
`b:lsbuffer_ignores` and `g:lsbuffer_ignores`. `z` is remapped to `Z`
and now toggles `b:lsbuffer_ignores` instead of the removed `s:ignores`.
These two lsbuffer_ignores are set up to allow for easy future additions
and user scriptability.

`lsbuffer#ls()` now centers the line with `:norm! zz`. Not very
customizable, but vim keeps putting the cursor on the last line at the
top of the screen with only one line visible because Bram is a dum-dum.

/plugin/lsbuffer.vim has been cleaned up quite a lot and there are no
longer `<plug>(...)`s for defaults, since `hasmapto()` works for any
*RHS* and the "plugs" provide nothing useful.

Lastly, `autocmd User LsBufferNewPre` has been added to trigger a User
autocommand just before the buffer-local maps are added to the lsbuffer.