better doc, hasmapto('Ls') normal mode
Listings are now indented in the help doc. This provides minor contrast
that is easier on the eys.

All of the global mappings are added now. I forgot to add them in
before.

Commands that can't use > at end of previous line due to indentation are
now wrapped in backticks to make them further distinguishable as
commands.

In plugin/, hasmapto() now checks normal mode, whereas before it was
checking every mode. Also, since it only checks substrings,
hasmapto('Lsnew') has been removed since it is redundant.