openwith, :Cd 0 or 1 arg
pressing `o` triggers input() with "open with: " prompt and completes
using shellcmd. If nothing entered, `xdg-open` is used. After pressing
<cr>, the external program or xdg-open is used. No error checking is
performed. This could be improved in a minor way by using
`executable()`.

`:Cd` is now equivalent to `:Cd ~` and the command itself has been
significantly improved, as now it uses <f-args> instead of <q-args> and
uses -nargs=? instead of -nargs=*. This prevents -complete=dir from
breaking and is more robust.

Since :Cd can produce zero arguments with <f-args>, lsbuffer#newcwd()'s
first argument, a:p for "path", is now optional and defaults to `~`
which is expand()ed to $HOME.